Super permissions
Added support for super permissions.

Super permissions allow a user of the library to express OR requirements for a given route. For example: 

We could have a route that requires the permission "team". However, using the super permission "admin" we can allow any user with the "admin" permission to access the route without explicitly holding the "team" permission.